Comment-Generator/requirements.txt

Flask==3.0.1
google-ai-generativelanguage==0.4.0
google-api-core==2.16.1
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
Werkzeug==3.0.1
wheel==0.37.1
